Everything builds!
Everybody knows that everything fails. But not every day.
Not today. Some days are special. Some days are so, so
blessed. Some days, nothing fails at all. Now and then,
Every once in a very long while, every day in a million days,
when the wind stands fair and the Doctor comes to call,
all tests pass.